+++ src/updates2mqtt/integrations/docker_enrich.py
@@ -533,7 +533,8 @@ class SourceReleaseEnricher:
                         api_response = alt_api_response
                     elif alt_api_results:
                         self.log.debug(
-                            "Failed to match latest release for %s, found tag %s for name %s",
+                            "Failed to match %s latest release for %s, found tag %s for name %s",
+                            registry_info.name,
                             detail.version,
                             alt_api_results.get("tag_name"),
                             alt_api_results.get("name"),